Рефакторинг

+++ lw2/primes/primes_utils.h
@@ -5,4 +5,4 @@ bool CheckArgumentCount(const int argumentCount);
 
 bool IsNumber(const std::string& str);
 
-std::set<uint32_t> GeneratePrimeNumbersSet(const unsigned upperBound);
+std::set<size_t> GeneratePrimeNumbersSet(const size_t upperBound);
